refactor(deepagents): extract prompt caching helpers
